+++ FaceDetectionComparison/face_detection_opencv_haar.py
@@ -1,4 +1,4 @@
-
+from __future__ import division
 import cv2
 import time
 import sys
@@ -27,7 +27,7 @@ def detectFaceOpenCVHaar(faceCascade, frame, inHeight=300, inWidth=0):
                   int(x2 * scaleWidth), int(y2 * scaleHeight)]
         bboxes.append(cvRect)
         cv2.rectangle(frameOpenCVHaar, (cvRect[0], cvRect[1]), (cvRect[2], cvRect[3]), (0, 255, 0),
-                      round(frameHeight / 150), 4)
+                      int(round(frameHeight / 150)), 4)
     return frameOpenCVHaar, bboxes
 
 if __name__ == "__main__" :
