+++ FaceDetectionComparison/run-all.py
@@ -1,4 +1,4 @@
-
+from __future__ import division
 import cv2
 import dlib
 import time
@@ -55,7 +55,7 @@ def detectFaceOpenCVHaar(faceCascade, frame, inHeight=300, inWidth=0):
                   int(x2 * scaleWidth), int(y2 * scaleHeight)]
         bboxes.append(cvRect)
         cv2.rectangle(frameOpenCVHaar, (cvRect[0], cvRect[1]), (cvRect[2], cvRect[3]), (0, 255, 0),
-                      round(frameHeight / 150), 4)
+                      int(round(frameHeight / 150)), 4)
     return frameOpenCVHaar, bboxes
 
 
@@ -77,7 +77,7 @@ def detectFaceOpenCVDnn(net, frame):
             x2 = int(detections[0, 0, i, 5] * frameWidth)
             y2 = int(detections[0, 0, i, 6] * frameHeight)
             bboxes.append([x1, y1, x2, y2])
-            cv2.rectangle(frameOpencvDnn, (x1, y1), (x2, y2), (0, 255, 0), round(frameHeight/150), 8)
+            cv2.rectangle(frameOpencvDnn, (x1, y1), (x2, y2), (0, 255, 0), int(round(frameHeight/150)), 8)
     return frameOpencvDnn, bboxes
 
 
@@ -102,7 +102,7 @@ def detectFaceDlibHog(detector, frame, inHeight=300, inWidth=0):
         cvRect = [int(faceRect.left()*scaleWidth), int(faceRect.top()*scaleHeight),
                   int(faceRect.right()*scaleWidth), int(faceRect.bottom()*scaleHeight) ]
         bboxes.append(cvRect)
-        cv2.rectangle(frameDlibHog, (cvRect[0], cvRect[1]), (cvRect[2], cvRect[3]), (0, 255, 0), round(frameHeight/150), 4)
+        cv2.rectangle(frameDlibHog, (cvRect[0], cvRect[1]), (cvRect[2], cvRect[3]), (0, 255, 0), int(round(frameHeight/150)), 4)
     return frameDlibHog, bboxes
 
 
@@ -127,7 +127,7 @@ def detectFaceDlibMMOD(detector, frame, inHeight=300, inWidth=0):
         cvRect = [int(faceRect.rect.left()*scaleWidth), int(faceRect.rect.top()*scaleHeight),
                   int(faceRect.rect.right()*scaleWidth), int(faceRect.rect.bottom()*scaleHeight) ]
         bboxes.append(cvRect)
-        cv2.rectangle(frameDlibMMOD, (cvRect[0], cvRect[1]), (cvRect[2], cvRect[3]), (0, 255, 0), round(frameHeight/150), 4)
+        cv2.rectangle(frameDlibMMOD, (cvRect[0], cvRect[1]), (cvRect[2], cvRect[3]), (0, 255, 0), int(round(frameHeight/150)), 4)
     return frameDlibMMOD, bboxes
